Move ResolverContext class to permit reuse
